+++ config.php
@@ -1,1 +1,9 @@
 <?php
+
+return [
+    'database' => [
+        'host' => '', // Database host
+        'dbname' => '', // Database name
+        'charset' => 'utf8mb4' // Character set for the database connection
+    ],
+];
