+++ controllers/index.php
@@ -1,5 +1,6 @@
 <?php
 
 view("index.phtml", [
+    'pageTitle' => 'Dashboard',
     'heading' => 'Home',
 ]);
