+++ models/Core/Database.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Core;
+namespace models\Core;
 use PDO;
 use PDOException;
