+++ models/Core/Validator.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace Core;
+namespace models\Core;
 
 class Validator
 {
+    // Validates a string based on its length
     public static function string($value, $min = 1, $max = INF)
     {
         $value = trim($value);
@@ -11,6 +12,7 @@ class Validator
         return strlen($value) >= $min && strlen($value) <= $max;
     }
 
+    // Validates whether an input is an email or not
     public static function email($value)
     {
         return filter_var($value, FILTER_VALIDATE_EMAIL);
