+++ models/Core/functions.php
