+++ models/DataSets/Company.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace DataSets;
+namespace models\DataSets;
 
 class Company
 {
