+++ models/DataSets/Course.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace DataSets;
+namespace models\DataSets;
 
 class Course
 {
