+++ models/DataSets/IndustriesDataSet.php
@@ -1,33 +1,33 @@
 <?php
 
-namespace DataSets;
+namespace models\DataSets;
 
 require_once('Core/Database.php');
-require_once('DataSets/UserType.php');
+require_once('DataSets/Industry.php');
 
-
-class UserTypesDataSet
+class IndustriesDataSet
 {
     protected $dbHandle, $dbInstance;
 
 
     public function __construct() {
-        $this->dbInstance = \Core\Database::getInstance();
+        $this->dbInstance = \models\Core\Database::getInstance();
         $this->dbHandle = $this->dbInstance->getdbConnection();
     }
 
-    public function fetchAllUserTypes()
+    public function fetchAllIndustries()
     {
-        $sqlQuery = 'SELECT * FROM userTypes';
+        $sqlQuery = 'SELECT * FROM industry';
 
         $statement = $this->dbHandle->prepare($sqlQuery); // prepare a PDO statement
         $statement->execute(); // execute the PDO statement
 
         $dataSet = [];
         while ($row = $statement->fetch()) {
-            $dataSet[] = new UserType($row);
+            $dataSet[] = new Industry($row);
         }
         return $dataSet;
 
     }
+
 }
