+++ models/DataSets/UserTypesDataSet.php
@@ -1,32 +1,33 @@
 <?php
 
-namespace DataSets;
-require_once('Core/Database.php');
-require_once('DataSets/Company.php');
+namespace models\DataSets;
 
-class CompaniesDataSets
+require_once('Core/Database.php');
+require_once('DataSets/UserType.php');
+
+
+class UserTypesDataSet
 {
     protected $dbHandle, $dbInstance;
 
 
     public function __construct() {
-        $this->dbInstance = \Core\Database::getInstance();
+        $this->dbInstance = \models\Core\Database::getInstance();
         $this->dbHandle = $this->dbInstance->getdbConnection();
     }
 
-    public function fetchAllCompanies()
+    public function fetchAllUserTypes()
     {
-        $sqlQuery = 'SELECT * FROM company';
+        $sqlQuery = 'SELECT * FROM userTypes';
 
         $statement = $this->dbHandle->prepare($sqlQuery); // prepare a PDO statement
         $statement->execute(); // execute the PDO statement
 
         $dataSet = [];
         while ($row = $statement->fetch()) {
-            $dataSet[] = new Company($row);
+            $dataSet[] = new UserType($row);
         }
         return $dataSet;
 
     }
-
 }
