+++ public/index.php
@@ -1,12 +1,12 @@
 <?php
-
+// Starting a PHP session
 session_start();
 
 // Define a constant for the base path directory for the project
 const BASE_PATH = __DIR__ . '/../';
 
 // Require necessary files
-require BASE_PATH . 'Core/functions.php';
+require BASE_PATH . 'models/Core/functions.php'; // Including functions file
 
 // Register an autoloader function
 spl_autoload_register(function ($class) {
@@ -15,11 +15,14 @@ spl_autoload_register(function ($class) {
     require base_path("{$class}.php");
 });
 
-$router = new \Core\Router();
+// Requiring and initializing an instance of Router to route user
+$router = new \models\Core\Router();
 $routes = require_once(base_path("routes.php"));
 
+// Parsing the requested URI and determining the request method
 $uri = parse_url($_SERVER['REQUEST_URI'])['path']; // Extract the path component from the current request's URI
 $method = $_POST['_method'] ?? $_SERVER['REQUEST_METHOD'];
 
+// Routing the URI and method
 $router->route($uri, $method);
