+++ views/index.phtml
@@ -1,1 +1,7 @@
+<?php require_once('partials/head.phtml') ?>
+<?php require_once('partials/nav.phtml') ?>
+<?php require_once('partials/header.phtml') ?>
+
 <p>BEST WEBSITE 2024</p>
+
+<?php require_once('partials/footer.phtml') ?>
