--- views/partials/header.phtml
@@ -1,1 +1,0 @@
-footer.phtml
